requirements.txt
python-decouple==3.8
pandas==2.3.3
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11